// src/redux/types.tsx
export enum Actions {
	SET_AUTHENTICATED = "SET_AUTHENTICATED",
	SET_UNAUTHENTICATED = "SET_UNAUTHENTICATED",
	SET_USER = "SET_USER",
	LOADING_USER = "LOADING_USER",
	MARK_NOTIFICATIONS_READ = "MARK_NOTIFICATIONS_READ",
	SAVE_POST = "SAVE_POST",
	UNSAVE_POST = "UNSAVE_POST",
	SET_ERRORS = "SET_ERRORS",
	LOADING_UI = "LOADING_UI",
	CLEAR_ERRORS = "CLEAR_ERRORS",
	LOADING_DATA = "LOADING_DATA",
	STOP_LOADING_UI = "STOP_LOADING_UI",
	SET_SAVED_POSTS = "SET_SAVED_POSTS",
	SET_POSTS = "SET_POSTS",
	SET_POST = "SET_POST",
	SET_REPLIES = "SET_REPLIES",
	POST_REPLY = "POST_REPLY",
	LIKE_POST = "LIKE_POST",
	UNLIKE_POST = "UNLIKE_POST",
	DELETE_POST = "DELETE_POST",
	POST_POST = "POST_POST",
	SUBMIT_COMMENT = "SUBMIT_COMMENT",
	EDIT_COMMENT = "EDIT_COMMENT",
	EDIT_POST = "EDIT_POST",
	DELETE_COMMENT = "DELETE_COMMENT",
	DELETE_REPLY = "DELETE_REPLY",
	EDIT_REPLY = "EDIT_REPLY",
}